Keep track of page numbers between refreshes
On a page of search results, we're tracking the current page number in
memory to know where we are. We also push that value into the browser
history under the query param `search_page`.

However, when a user browses away and comes back, we aren't passing that
query parameter through from the controller to the template that
renders the page, so we're just starting again each time.

A better improvement might be to rework the Javascript code itself to
use the query parameter rather than rely on it being rendered at the PHP
level.