Render author results on page load
Pre-load the first page of results, this saves the browser having to
make a second request. I'm conflicted about this though. I think there
are two possible jobs-to-be-done when people land on the Author page:

 * "I want to know more about the Author"
 * "I want to see other works by the Author"

If the point is to know more about the author, then this change is bad
-- it delays loading the profile while we fetch the first page of
results.

If the point is to see other works by the author, then this change is
good -- we've elimitated a round-trip where we had to load the profile
and _then_ load the books by that author.